Refactoring for sonarQube
