+++ tests/bsv/auth/test_auth_peer_basic.py
@@ -8,46 +8,7 @@ from bsv.auth.peer_session import PeerSession
 from bsv.auth.session_manager import DefaultSessionManager
 from bsv.keys import PrivateKey
 
-
-class LocalTransport:
-    def __init__(self):
-        self._on_data_callback = None
-        self.sent_messages: list[AuthMessage] = []
-
-    def on_data(self, callback):
-        self._on_data_callback = callback
-        return None
-
-    def send(self, message_or_ctx, message=None):
-        # Handle both calling patterns:
-        # - send(message) - peer.py calls it this way
-        # - send(ctx, message) - interface defines it this way
-        if message is None:
-            # Called as send(message) - first arg is the message
-            msg = message_or_ctx
-        else:
-            # Called as send(ctx, message) - first arg is ctx, second is message
-            msg = message
-        self.sent_messages.append(msg)
-        # Note: peer.py callback expects just (message), not (ctx, message)
-        if self._on_data_callback is not None:
-            return self._on_data_callback(msg)
-        return None
-
-
-class GetPub:
-    def __init__(self, pk):
-        self.public_key = pk
-
-
-class Sig:
-    def __init__(self, signature: bytes):
-        self.signature = signature
-
-
-class Ver:
-    def __init__(self, valid: bool):
-        self.valid = valid
+from .conftest import LocalTransport, GetPub, Sig, Ver
 
 
 class MockWallet:
